removing unused jars, trying to resolve JSON in client

+++ service/src/main/resources/sql/delete-data.sql
@@ -1,3 +1,3 @@
-delete from UserArticle;
+delete from UserArticle where userid=2;
 delete from Article;
 delete from SectionArticle;
